Further simplfy serialized suite consumer test
