+++ tests/Functional/RemoteEvent/SourcesSerializedSuiteStateChangedWebhookConsumer/NoEventDispatchedTest.php
@@ -5,43 +5,16 @@ declare(strict_types=1);
 namespace App\Tests\Functional\RemoteEvent\SourcesSerializedSuiteStateChangedWebhookConsumer;
 
 use App\Event\SerializedSuiteRetrievedEvent;
-use App\Tests\Application\AbstractApplicationTest;
-use App\Tests\Functional\Application\GetClientAdapterTrait;
-use App\Tests\Services\EventSubscriber\EventRecorder;
-use App\Tests\Services\Factory\RemoteEventConfigurationFactory;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Symfony\Component\RemoteEvent\RemoteEvent;
 use Symfony\Component\Uid\Ulid;
 
-class NoEventDispatchedTest extends AbstractApplicationTest
+class NoEventDispatchedTest extends AbstractConsumerTestCase
 {
-    use GetClientAdapterTrait;
-
-    private EventRecorder $eventRecorder;
-    private RemoteEventConfigurationFactory $remoteEventConfigurationFactory;
-    private string $sourcesNotifySecret;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $eventRecorder = self::getContainer()->get(EventRecorder::class);
-        \assert($eventRecorder instanceof EventRecorder);
-        $this->eventRecorder = $eventRecorder;
-
-        $remoteEventConfigurationFactory = self::getContainer()->get(RemoteEventConfigurationFactory::class);
-        \assert($remoteEventConfigurationFactory instanceof RemoteEventConfigurationFactory);
-        $this->remoteEventConfigurationFactory = $remoteEventConfigurationFactory;
-
-        $sourcesNotifySecret = self::getContainer()->getParameter('sources_notify_secret');
-        \assert(is_string($sourcesNotifySecret));
-        $this->sourcesNotifySecret = $sourcesNotifySecret;
-    }
-
     #[DataProvider('noEventIsDispatchedDataProvider')]
     public function testNoEventIsDispatched(RemoteEvent $event): void
     {
-        $remoteEventConfiguration = $this->remoteEventConfigurationFactory->create($event, $this->sourcesNotifySecret);
+        $remoteEventConfiguration = $this->remoteEventConfigurationFactory->create($event, $this->notifySecret);
 
         $response = self::$staticApplicationClient->makeSourcesSerializedSuiteStateChangedNotifyRequest(
             $remoteEventConfiguration->headers,
